requirements.txt
requests
python-dotenv
openai
tiktoken
tqdm
playwright
